+++ src/index.common.ts
@@ -22,4 +22,4 @@ export * as Models from './models'
 export * as SPDX from './spdx'
 export * as Spec from './spec'
 export * as Types from './types'
-// do not export the helpers, they are for internal use only
+// do not export the _helpers, they are for internal use only
