+++ src/models/externalReference.ts
@@ -17,8 +17,8 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
+import { Comparable, SortableSet } from '../_helpers/sortableSet'
 import { ExternalReferenceType } from '../enums'
-import { Comparable, SortableSet } from '../helpers/sortableSet'
 
 interface OptionalProperties {
   comment?: ExternalReference['comment']
